third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-type-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'type' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matrixElement.getAttribute('type') is "saturate"
PASS saturateElement.getAttribute('type') is "hueRotate"
PASS hueRotateElement.getAttribute('type') is "luminanceToAlpha"
PASS luminanceToAlphaElement.getAttribute('type') is "matrix"
PASS matrixElement.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS saturateElement.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE
PASS hueRotateElement.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE
PASS luminanceToAlphaElement.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA
PASS successfullyParsed is true

TEST COMPLETE
